refacto test
