+++ product/models.py
@@ -7,14 +7,6 @@ from user.models import Profile
 from tools import logger
 
 
-# class Profile(models.Model):
-#     """
-#     models profile for add user image
-#     """
-#     user = models.OneToOneField(User, on_delete=models.CASCADE)
-#     image = models.BinaryField(null=True)
-
-
 class Category(models.Model):
     """
     models category for product
